third_party/blink/web_tests/external/wpt/css/css-anchor-position/scrollable-containing-block-size-expected.txt

This is a testharness.js-based test.
[FAIL] .target 1
  assert_equals: \n<div class="target" data-offset-x="0" data-offset-y="0" data-expected-width="200" data-expected-height="200"></div>\nwidth expected 200 but got 100
[FAIL] .target 2
  assert_equals: \n<div class="target" data-offset-x="0" data-offset-y="0" data-expected-width="200" data-expected-height="200"></div>\nwidth expected 200 but got 100
[FAIL] .target 3
  assert_equals: \n<div class="target" data-offset-x="0" data-offset-y="0" data-expected-width="200" data-expected-height="200"></div>\nwidth expected 200 but got 100
[FAIL] .target 4
  assert_equals: \n<div class="target" data-offset-x="0" data-offset-y="0" data-expected-width="200" data-expected-height="200"></div>\nwidth expected 200 but got 100
[FAIL] .target 5
  assert_equals: \n<div class="target" data-offset-x="0" data-offset-y="0" data-expected-width="200" data-expected-height="200"></div>\nwidth expected 200 but got 100
[FAIL] .target 6
  assert_equals: \n<div class="target" data-offset-x="0" data-offset-y="0" data-expected-width="200" data-expected-height="200"></div>\nwidth expected 200 but got 100
[FAIL] .target 7
  assert_equals: \n<div class="target" data-offset-x="0" data-offset-y="0" data-expected-width="200" data-expected-height="200"></div>\nwidth expected 200 but got 100
[FAIL] .target 8
  assert_equals: \n<div class="target" data-offset-x="0" data-offset-y="0" data-expected-width="200" data-expected-height="200"></div>\nwidth expected 200 but got 100
[FAIL] .target 9
  assert_equals: \n<div class="target" data-offset-x="0" data-offset-y="0" data-expected-width="200" data-expected-height="200"></div>\nwidth expected 200 but got 100
[FAIL] .target 10
  assert_equals: \n<div class="target" data-offset-x="0" data-offset-y="0" data-expected-width="200" data-expected-height="200"></div>\nwidth expected 200 but got 100
[FAIL] .target 11
  assert_equals: \n<div class="target" data-offset-x="0" data-offset-y="-100" data-expected-width="200" data-expected-height="200"></div>\nwidth expected 200 but got 100
[FAIL] .target 12
  assert_equals: \n<div class="target" data-offset-x="-100" data-offset-y="0" data-expected-width="200" data-expected-height="200"></div>\nwidth expected 200 but got 100
[FAIL] .target 13
  assert_equals: \n<div class="target" data-offset-x="-100" data-offset-y="0" data-expected-width="200" data-expected-height="200"></div>\nwidth expected 200 but got 100
Harness: the test ran to completion.
